requirements.txt
Django
djangorestframework
Pillow
django-cors-headers
djangorestframework-recursive
django-extensions
django-bootstrap5
django-multiselectfield
django-registration